Create Save Results functionality
Related to Issue #4: https://github.com/alyssonfm/contractok/issues/4

The button of Save results was plotted, but no functionality was
implemented. Now, the button open a Dialog for user, to select a folder,
then copying temp/ContractOk/result files to it. I have none acknowledge
of any side-effects for this.

I'll work next on printing a dialog confirming the action for the user.