Create confirm message dialog for Save Results
Related to #4

The button of Save Results, if clicked didn't show a message confirming
sucess of the action to user. Now, I've used a MessageBox dialog to
print message confirming sucess.

We have to study a little more the code, to see if the failure message
will be shown to user.