Update Nuget Packages and Randoop
The Randoop version, wasn't working on one build of Windows 10 that I
was using. That because Randoop wasn't eliminating '.' on miliseconds
string when generating RandoopTest filename. So I put to eliminate too.

Also the NuGet packages needed an update since Visual Studio 2015 it's
on 1.0 version. Then Roslyn project had finished with a 1.0 build, and
for that I need to update it.